+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Changelog
 - Added a link to said documentation in the settings menu.
 - Added a link to show the Options Page in the popup.
 - Fixes *Options page* refreshing after selecting an option in the *General* category.
+- Fixes *Options page* refreshing after (un)checking the secondary background checkbox in 
+  `druthers`.
+- Fixes secondary background checkbox visibility when selecting a theme from a dropdown menu in 
+  `druthers`.
 
 # 18.0.0 [Device Theme Sync]